+++ src/reducers/index.js
@@ -185,6 +185,8 @@ export const getNewSearchShown = (state) => fromPrefs.getNewSearchShown(state.pr
 
 export const getUseNewSearch = (state) => fromPrefs.getUseNewSearch(state.prefs);
 
+export const getPinnedQueries = (state) => fromPrefs.getPinnedQueries(state.prefs);
+
 export const getOptionList = (state, optionListName) => (
   fromOptionList.get(state.optionList, optionListName)
 );